FastAPI-App/requirements.txt
fastapi[all]
pydantic
uvicorn[standard]
python-multipart
numpy
Pillow
tensorflow
PyJWT